pyproject.toml
[project]
name = "comfyui-agent"
version = "0.1.0"
description = "AI Influencer Studio — unified backend for trend parsing, video generation, and bot"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "requests>=2.28",
    "websocket-client>=1.5",
    "huggingface_hub>=0.20",
    "python-telegram-bot>=21.0",
    "httpx>=0.27",
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
vps = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.20",
    "python-dateutil>=2.8",
    "instaloader>=4.10",
    "TikTokApi>=6.0",
    "playwright",
    "yt-dlp",
]

[project.scripts]
comfy-pipeline = "comfy_pipeline.cli:main"
vast-agent = "vast_agent.cli:main"
comfy-bot = "telegram_bot.bot:main"
isp-pipeline = "isp_pipeline.cli:main"
comfy-api = "api.app:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
